refactor: standardize import paths & reorganize component structure
