+++ src/sections/Footer.tsx
@@ -1,7 +1,7 @@
 // src/sections/Footer.tsx
 
 // footer component
-import { SocialIcons } from "../components/ui/SocialIcons";
+import { SocialIcons } from "~/components/ui/SocialIcons";
 
 
 // footer section
