+++ src/sections/home/HeroSection.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
-import FloatingParticle from '~/components/FloatingParticle';
+import FloatingParticle from '~/components/ui/FloatingParticle';
 import useTypingAnimation from '~/hooks/useTypingAnimation';
 interface HeroSectionProps {
   scrollToSection: (sectionId: string) => void;
